widget intent now jumps to matching day and opens match smooth scrolling is still missing reformated match_id from double to long (currently int at api)

+++ app/src/main/java/barqsoft/footballscores/ViewHolder.java
@@ -14,7 +14,7 @@ public class ViewHolder {
     public TextView date;
     public ImageView homeCrest;
     public ImageView awayCrest;
-    public double matchId;
+    public long matchId;
 
     public ViewHolder(View view) {
         homeName = (TextView) view.findViewById(R.id.home_name);
